--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -8,6 +8,7 @@ using DataFrames
 using CSV
 
 const S = ScientificTypes
+const M = S.MLJ()
 
 include("type_tests.jl")