--- a/test/type_tests.jl
+++ b/test/type_tests.jl
@@ -39,8 +39,8 @@ end
 
 @testset "scitype of a table that is also an array" begin
     X = (x=rand(4),)
-    f = mkpath()
-    CSV.write(f, X)
-    file = CSV.file(f)
+    CSV.write("test.csv", X)
+    file = CSV.file("test.csv")
     @test scitype(file) == scitype(X)
+    rm("test.csv")
 end