* Add `Textual` (see #72 )
* Further performance improvements (see #70 )

For 2, among other things there's a better treatment for missing values, before we were doing:

```julia
vtight = broadcast(identity, v)
if typeof(vtight) = typeof(v)
```

but actually it's much faster to just check if there are true missings in `v`. 

A few other improvements here and there, results show a nice speedup:

```julia
using ScientificTypes, DataFrames

function foo()
    r = rand()
    r < 0.2 && return "AAA"
    r < 0.4 && return "BBB"
    r < 0.6 && return "CCC"
    r < 0.8 && return "DDD"
    return "EEE"
end

v1 = [foo() for i in 1:1_000_000];
v2 = [foo() for i in 1:1_000_000];
df = DataFrame((x1=v1,x2=v2));

using BenchmarkTools
using CategoricalArrays

@btime categorical($v1); # 28.450 ms (1000046 allocations: 34.34 MiB)
@btime coerce($v1, Multiclass);
#this branch ==> 45.596 ms (1000080 allocations: 31.48 MiB)
#master ==> 215.418 ms (1000086 allocations: 39.11 MiB)

function foo()
    r = rand()
    r < 0.2 && return "AAA"
    r < 0.4 && return "BBB"
    r < 0.6 && return "CCC"
    r < 0.8 && return "DDD"
    return missing
end

v2 = [foo() for i in 1:1_000_000];

@btime categorical($v2); # 25.279 ms (799758 allocations: 28.22 MiB)

@btime coerce($v2, Union{Missing,Multiclass});
#this branch ==> 139.077 ms (1599618 allocations: 51.68 MiB)
#master ==> 380.969 ms (1799840 allocations: 71.14 MiB)

using Random, Distributions

n   = 1_115_000
df  = DataFrame(
        rs  = rand(["A", "B", "C", "D", "E", "F", "G"], n),
        es  = rand(["1","3","5","7","8","9"], n),
        inc = rand([collect(5000:15000); [missing for _ ∈ 1:50]], n),
        gr  = rand([randstring(6) for _ in 1:400], n),
        year = rand(2012:2017, n),
        lt = rand([12, 24, 36, 48, 60, 72], n),
        pb = Int.(round.(rand(LogNormal(8.07, 1.116), n))),
        cb = rand(LogNormal(7.88, 1.288), n),
        ci = rand(0.03:0.01:0.17, n),
        db = rand([repeat(collect(1945:2000),100); missing], n),
        ms = rand(["1", "2", "NA"], n),
        ho = BitArray(rand([true, false], n)),
        el = Int.(round.(rand(LogNormal(1.0, 0.72), n))),
        fl = rand(n),
        ft = rand(n),
        ea = BitArray(rand([true, false], n))
        );

@btime schema($df);
#this branch ==> 249.195 μs (140 allocations: 10.78 KiB)
#master ==> 7.925 ms (158 allocations: 36.16 MiB)

@btime coerce($df, autotype($df, :few_to_finite));
#this branch ==> 1.503 s (12282563 allocations: 609.48 MiB)
#master ==> 3.985 s (12263782 allocations: 756.95 MiB)
```

## Notes

There's still roughly a factor 2 performance drop against `categorical` in the non-missing case which we should be able to reduce. In the case of missings there's the additional cost of providing the "tight" version that we put on ourselves.